GH-57 - Introduce @Identity in DDD annotations.
Although not an original concept within the DDD book, an identifier is already part of the type based domain language in the DDD module. For symmetry reasons it's reasonable to also allow to express the concept as annotation. We went for Identity (over Identifier) for the following reason the annotation is used on a field declaration. The field is not the identifier itself but the value held in the field. I.e. the field's value constitutes the identity of the surrounding type.