GH-93 - Introduce CQRS annotations in cqrs package.
Introduce a copy of the annotations currenly available in cqrs.annotation in the cqrs package. This is to create symmetry between all architectural jMolecules abstractions. The legacy annotations are deprecated for removal in 2.0 and annotated with the new ones in turn so that processing technology can look out for the new ones only and still find the legacy ones.