GH-152 - Introduce @Stereotype meta-annotation.
We now ship an @Stereotype annotation in a jStereotype artifact that allows marking annotations and types as stereotypes. An API working with that metadata will subsequently be introduced in jMolecules Integrations.

The jMolecules DDD, jMolecules Onion and Hexagonal Architecture artifacts have been fully enriched with both the annotations and the derived JSON metadata.