GH-152 - Introduce @Stereotype meta-annotation.
We now ship an @Stereotype annotation in a jstereotype artifact that allows marking annotations and types as stereotypes. An API working with that metadata will subsequently be introduced in jMolecules Integrations.

The jmolecules-ddd and jmolecules-hexagonal-architecture artifact have been fully enriched with both the annotations and the derived JSON metadata. Other artifacts have been annotated with @Stereotype but not yet fully described in terms of stereotype hierarchies.